requirements.txt
anthropic>=0.40.0
supabase>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
apscheduler>=3.10.0
